+++ recording/lib/storage/file.ex
@@ -5,15 +5,15 @@ defmodule Membrane.RTC.Engine.Endpoint.Recording.Storage.File do
   @behaviour Membrane.RTC.Engine.Endpoint.Recording.Storage
 
   @impl true
-  def get_sink(config, _storage_opts) do
-    location = Path.join(config.path_prefix, config.filename)
+  def get_sink(config, %{output_dir: output_dir}) do
+    location = Path.join(output_dir, config.filename)
     File.touch!(location)
     %Membrane.File.Sink{location: location}
   end
 
   @impl true
-  def save_object(config, _storage_opts) do
-    location = Path.join(config.path_prefix, config.filename)
+  def save_object(config, %{output_dir: output_dir}) do
+    location = Path.join(output_dir, config.filename)
 
     with :ok <- File.touch(location) do
       File.write(location, config.object)