+++ recording/lib/storage/s3.ex
@@ -16,12 +16,13 @@ defmodule Membrane.RTC.Engine.Endpoint.Recording.Storage.S3 do
           bucket: String.t()
         }
 
-  @type storage_opts :: %{credentials: credentials_t()}
+  @type storage_opts :: %{:credentials => credentials_t(), optional(:path_prefix) => Path.t()}
 
   @impl true
   @spec get_sink(Storage.recording_config(), storage_opts()) :: struct()
   def get_sink(config, storage_opts) do
-    path = Path.join(config.path_prefix, config.filename)
+    path_prefix = Map.get(storage_opts, :path_prefix, "")
+    path = Path.join([path_prefix, config.recording_id, config.filename])
 
     %__MODULE__.Sink{
       path: path,
@@ -32,7 +33,8 @@ defmodule Membrane.RTC.Engine.Endpoint.Recording.Storage.S3 do
 
   @impl true
   def save_object(config, storage_opts) do
-    path = Path.join(config.path_prefix, config.filename)
+    path_prefix = Map.get(storage_opts, :path_prefix, "")
+    path = Path.join([path_prefix, config.recording_id, config.filename])
     credentials = storage_opts.credentials
     aws_config = create_aws_config(credentials)