+++ recording/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.1.0-dev
-* Add working AWS storage [#357](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/360)
 * Initial release [#356](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/356)
 * Add crash groups, fix adding nonexistent tracks to state, fix READMEs [#366](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/366)
+* Add working AWS storage [#369](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/369)
+* Refactor storage API [#375](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/375)